Feature/outsource contributing guideline (#25)
Create CONTRIBUTING.md

README.md contains all information from usage of the chaos-kit and its contribution guideline. When Azure pipelines generate the docs for the docs.proofdock.io the contribution guidelines are irrelevant. Sourcing those out to a seperate file contributes to a eased readability. We did it already for the chaos-middleware.